requirements.txt
boto3==1.16.63
botocore==1.19.63
click==7.1.2
colorclass==2.2.0
jmespath==0.10.0
python-dateutil==2.8.1
PyYAML==5.4.1
s3transfer==0.3.4
six==1.15.0
terminaltables==3.1.0
urllib3==1.26.3
XlsxWriter==1.3.7
